Fix (FAPI & FreeType)
Bug #691284 "Error: /unknownerror in --.FAPIBuildChar"
The issue is caused because the PostScript file adds to an existing type 1 font a
CharString entry which is not in type 1 CharString format. I would have thought this
to be illegal but other PostScript interpreters (specifically Distiller) are happy
with it.

Modifed the FAPI interface so that after retrieving a type 1, non-CID, CharString we
test to see if it is a procedure, if it is, then we return an error from the FAPI
code to the caller (in this case the FreeType interface).

Modified the FreeType interface so that if the CharString retrieval returns an error
we return a  specific (otherwise unused) error to the caller, in this case the
FreeType engine. On return from FreeType to the interface we test for the specific
FreeType error (FT_ERR_Invalid_File_Format) and if this is detected then instead of
returning an error to FAPI, we add one to the character index (to ensure it isn't 0)
and return that to FAPI.

Finally, in the FAPI code, after 'rendering' the glyph, if we were returned a value
> 0 then we assume that 'something went wrong' with the CharString. As a last-ditch
effort we use the return value -1 to retrieve the CharString, and then 'exec' it.

NB if we fail to retrieve the CharString at this level, we return an invalidfont error.

Expected Differences
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11191 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/psi/ifapi.h
@@ -126,7 +126,7 @@ struct FAPI_font_s {
     int		   (*get_proc) (FAPI_font *ff, fapi_font_feature var_id, int index, char *Buffer);
     unsigned short (*get_gsubr)(FAPI_font *ff, int index,     byte *buf, ushort buf_length);
     unsigned short (*get_subr) (FAPI_font *ff, int index,     byte *buf, ushort buf_length);
-    unsigned short (*get_glyph)(FAPI_font *ff, int char_code, byte *buf, ushort buf_length);
+    int		   (*get_glyph)(FAPI_font *ff, int char_code, byte *buf, ushort buf_length);
     unsigned short (*serialize_tt_font)(FAPI_font *ff, void *buf, int buf_size);
 };
